move all completions code to jupyter_ai.completions
