+++ packages/jupyter-ai/jupyter_ai/completions/completion_types.py
