--- packages/jupyter-ai/jupyter_ai/completions/models.py
@@ -1,17 +1,0 @@
-from jupyter_ai_magics.models.completion import (
-    CompletionError,
-    InlineCompletionItem,
-    InlineCompletionList,
-    InlineCompletionReply,
-    InlineCompletionRequest,
-    InlineCompletionStreamChunk,
-)
-
-__all__ = [
-    "InlineCompletionRequest",
-    "InlineCompletionItem",
-    "CompletionError",
-    "InlineCompletionList",
-    "InlineCompletionReply",
-    "InlineCompletionStreamChunk",
-]
